+++ src/main/java/com/buildershandbag/integration/ArchitectureCraftIntegration.java
@@ -17,10 +17,6 @@ final class ArchitectureCraftIntegration {
     private ArchitectureCraftIntegration() {
     }
 
-    // FIXME: *Sigh* ArchitectureCraft seems to be suffering from the same problem as BlockCraftery :
-    //        the rendering may be stale until neighbor updates occur.
-    //        Need the same workaround (PacketBlockcrafteryTileSync packet after placing)
-
     // TODO: If there is demand, we might consider fractional material cost
     //       (aligning with ArchitectureCraft's own cost calculations),
     //       which would require custom rendering, decoupling ItemStack